starter/requirements.txt
psycopg2==2.7.3
celery==4.1.0
django-crispy-forms==1.6.1
django-polymorphic==1.2
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.3.1
django-rest-auth==0.9.1
django-bootstrap-form==3.2.1
django-allauth-bootstrap==0.2 #Django==1.9.3
django-countries==4.6.1
django-location-field==2.0.3
django-phonenumber-field==1.3.0
django-organizations==0.9.2
django-taggit==0.22.1
django-taggit-helpers==0.1.4
django_ddp==0.19.1
django-qrcode==0.3
django-report-builder==3.5.0
django-cors-headers==2.1.0
django_admin_honeypot==1.0.0
django-changuito==1.0
django-localflavor==1.5.2
django-dashing==0.3.3
django-stored-messages==1.4.0
django_comments_xtd==2.0.5
django-bower==5.2.0
schedule==0.4.3
django-geojson==2.10.0
django-mptt==0.8.7
django-treenav==1.1.0
django-channels==0.7.0
bootstrap-admin==0.3.7.1
django-taggit-labels==0.5.1
django-autoslug==1.9.3
pygments==2.2.0
django-model-utils==3.0.0
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.11.0
django-leaflet==0.22.0

#patterns==0.3
django-reversion==2.0.9
#django-haystack==2.6.1
#django-messages==0.5.3

#django_push_notifications==1.5.0 # Django==1.9.3
#django-rest-framework-docs==0.1.7 # Django<1.10
#xadmin==0.6.1
#django-xadmin==0.5.0 # import imp/importlib importlib/imp.reload() instead of reload()

django-storages==1.6.5
boto3==1.4.5

coreapi==2.3.1
markdown==2.6.8
drfdocs==0.0.11
django-filter==1.0.4

django-rest-raml==0.1.0
raml-codec==0.1.4

django-rest-swagger==2.1.2

#python 3.6.2
uwsgi==2.0.15
django==1.9.9
django-test-without-migrations==0.6
